refactor - folder structure updated and absolute path added
