+++ src/components/Character/character.module.css
@@ -1,12 +1,3 @@
-.pgHeader {
-  border-bottom: solid thin var(--clr-off-primary);
-}
-
-.pgNav {
-  width: 60vw;
-  overflow-x: auto;
-}
-
 .cardsGrid {
   gap: 1rem;
   place-content: center;
@@ -21,7 +12,7 @@
   width: 20rem;
 }
 
-.charImage {
+.charImg {
   width: 100%;
   object-fit: cover;
 }
