+++ src/components/Character/index.js
@@ -1,0 +1,2 @@
+export { default as CharacterCard } from './CharacterCard'
+export { default as CharacterModal } from './CharacterModal'
