+++ src/components/Header/header.module.css
@@ -1,0 +1,3 @@
+.pgHeader {
+  border-bottom: solid thin var(--clr-off-primary);
+}
