+++ src/components/Nav/index.js
@@ -1,0 +1,1 @@
+export { default as PaginationNav } from './PaginationNav'
