+++ src/components/Nav/nav.module.css
@@ -1,0 +1,4 @@
+.pgNav {
+  width: 60vw;
+  overflow-x: auto;
+}
