+++ src/contexts/index.js
@@ -1,0 +1,2 @@
+export { useModal, ModalProvider } from './modal.context'
+export { usePage, PageProvider } from './page.context'
