+++ src/index.js
@@ -1,8 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
-import { ModalProvider } from "./contexts/modal.context";
-import { PageProvider } from "./contexts/page.context";
+import { ModalProvider, PageProvider } from "contexts";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
